refactor: 게시글 관련 repository 파일 패키징 변경
