+++ src/main/java/com/flint/flint/community/repository/post/PostRepository.java
@@ -1,9 +1,7 @@
-package com.flint.flint.community.repository;
+package com.flint.flint.community.repository.post;
 
 import com.flint.flint.community.domain.post.Post;
-import com.flint.flint.community.repository.custom.PostRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 
 public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
 }
