+++ src/main/java/com/flint/flint/community/repository/post/PostRepositoryCustom.java
@@ -1,15 +1,15 @@
-package com.flint.flint.community.repository.custom;
+package com.flint.flint.community.repository.post;
 
 import com.flint.flint.community.domain.post.Post;
+import com.flint.flint.community.dto.response.PostListResponse;
+import com.flint.flint.community.spec.SortStrategy;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.util.List;
 
-/**
- * @author 정순원
- * @since 2023-11-06
- */
 public interface PostRepositoryCustom {
+    List<PostListResponse> findPostsWithBoardByPaging(Long boardId, Post cursorPost, Long size, SortStrategy sort);
 
     /**
      * 전체게시판에서
