+++ src/main/java/com/flint/flint/community/repository/post_comment/PostCommentRepository.java
@@ -1,4 +1,4 @@
-package com.flint.flint.community.repository;
+package com.flint.flint.community.repository.post_comment;
 
 import com.flint.flint.community.domain.post.PostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
