+++ src/main/java/com/flint/flint/community/repository/post_comment/PostCommentRepositoryImpl.java
@@ -1,8 +1,6 @@
-package com.flint.flint.community.repository.impl;
+package com.flint.flint.community.repository.post_comment;
 
 import com.flint.flint.community.domain.post.PostComment;
-import com.flint.flint.community.domain.post.QPostComment;
-import com.flint.flint.community.repository.custom.PostCommentRepositoryCustom;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
