+++ src/test/java/com/flint/flint/community/controller/PostCommentUpdateControllerTest.java
@@ -3,7 +3,7 @@ package com.flint.flint.community.controller;
 import com.flint.flint.community.domain.post.PostComment;
 import com.flint.flint.community.domain.post.PostCommentLike;
 import com.flint.flint.community.repository.PostCommentLikeRepository;
-import com.flint.flint.community.repository.PostCommentRepository;
+import com.flint.flint.community.repository.post_comment.PostCommentRepository;
 import com.flint.flint.custom_member.WithMockCustomMember;
 import com.flint.flint.member.domain.main.Member;
 import com.flint.flint.member.repository.MemberRepository;
