+++ src/test/java/com/flint/flint/community/service/PostCommentUpdateServiceTest.java
@@ -6,8 +6,8 @@ import com.flint.flint.community.domain.post.Post;
 import com.flint.flint.community.domain.post.PostComment;
 import com.flint.flint.community.dto.request.PostCommentUpdateRequest;
 import com.flint.flint.community.dto.response.PostCommentUpdateResponse;
-import com.flint.flint.community.repository.PostCommentRepository;
-import com.flint.flint.community.repository.PostRepository;
+import com.flint.flint.community.repository.post_comment.PostCommentRepository;
+import com.flint.flint.community.repository.post.PostRepository;
 import com.flint.flint.member.domain.main.Member;
 import com.flint.flint.member.repository.MemberRepository;
 import com.flint.flint.member.service.MemberService;
