+++ src/test/java/com/flint/flint/community/service/PostServiceTest.java
@@ -9,7 +9,7 @@ import com.flint.flint.community.dto.request.PostRequest;
 import com.flint.flint.community.dto.response.PostListResponse;
 import com.flint.flint.community.dto.response.PostPreSignedUrlResponse;
 import com.flint.flint.community.repository.BoardRepository;
-import com.flint.flint.community.repository.PostRepository;
+import com.flint.flint.community.repository.post.PostRepository;
 import com.flint.flint.community.spec.BoardType;
 import com.flint.flint.community.spec.SortStrategy;
 import com.flint.flint.member.domain.main.Member;
